// src/app/pages/pets/interfaces/pet-dto.ts
export interface PetDto {
  id: number;
  name: string;//
  gender: string;//
  birthDate: string;////////
  registerDate: string;
  colour: string;////////
  specificBreed: null;//////////
  characteristic: string;////////
  size: string;////////
  species: string;//
  speciesId: string;//
  breed: string;//
  breedId: string;//
  owner_id: number;
  detail_id: number;
  encoded: string;
  state: number;
}